+++ vibetags/src/main/java/se/deversity/vibetags/processor/internal/content/annotations/AISecureLoggingFormatter.java
@@ -4,6 +4,7 @@ import javax.lang.model.element.Element;
 import se.deversity.vibetags.annotations.AISecureLogging;
 import se.deversity.vibetags.processor.internal.ElementNaming;
 import se.deversity.vibetags.processor.internal.content.AnnotationFormatter;
+import se.deversity.vibetags.processor.internal.content.Escape;
 import se.deversity.vibetags.processor.internal.content.Platform;
 
 /**
@@ -24,7 +25,7 @@ public final class AISecureLoggingFormatter implements AnnotationFormatter {
 
         switch (platform) {
             case CLAUDE:
-                sb.append("    <file path=\"").append(className).append("\">\n      <logging_policy>").append(policy.name()).append("</logging_policy>\n    </file>\n");
+                sb.append("    <file path=\"").append(Escape.xml(className)).append("\">\n      <logging_policy>").append(Escape.xml(policy.name())).append("</logging_policy>\n    </file>\n");
                 break;
             case LLMS_FULL:
                 sb.append("### ").append(className).append("\n- **Secure Logging Policy**: ").append(policy.name()).append("\n\n");